requirements.txt
streamlit
langchain>=0.1.17
langgraph
langchain-groq
python-dotenv
chromadb
huggingface-hub
sentence-transformers
pypdf